Rename loader to module-helpers
