+++ lib/module-helpers.js
