+++ lib/modules-view.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const Humanize = require('humanize-plus')
-const Loader = require('./loader')
+const modules = require('./module-helpers')
 const ModuleView = require('./module-view')
 const View = require('./view')
 
@@ -43,7 +43,7 @@ class ModulesView extends View {
   }
 
   loadGraph () {
-    Loader.getRenderModules().then((mainModule) => {
+    modules.getRenderModules().then((mainModule) => {
       const suffix = this.getTabLabelSuffix(mainModule)
       this.renderProcessTab.textContent = `Renderer Process ${suffix}`
       this.renderRequireRows.innerHTML = ''
@@ -54,7 +54,7 @@ class ModulesView extends View {
       console.error(error.stack || error)
     })
 
-    Loader.getMainModules().then((mainModule) => {
+    modules.getMainModules().then((mainModule) => {
       const suffix = this.getTabLabelSuffix(mainModule)
       this.mainProcessTab.textContent = `Main Process ${suffix}`
       this.mainRequireRows.innerHTML = ''
